Updated Column and moved types
