+++ query-builder/Table.ts
@@ -1,21 +1,6 @@
 import { Column } from "./Column.ts";
-import { ColumnTypes, typeMap, TypeMapEl } from "./TypeUtils.ts";
+import { ColumnTypes, typeMap, TypeMapEl, EnumColumn, TableConstraints } from "./TypeUtils.ts";
 import { DBDialects } from "../types.ts";
-
-export interface EnumColumn {
-  name: string;
-  columns: string[];
-}
-
-export interface TableConstraints {
-  unique: string[][];
-  primary?: string[];
-  index: string[];
-  enums: EnumColumn[];
-  updatedAt: boolean;
-  ifNotExists?: boolean;
-  isTemporary?: boolean;
-}
 
 /** The table class exposed in the second argument `schema.create()` method.
  * 
@@ -98,9 +83,9 @@ export class Table {
       default:
         return `CREATE${
           this.constraints.isTemporary ? " TEMPORARY" : ""
-        } TABLE${
+          } TABLE${
           this.constraints.ifNotExists ? " IF NOT EXISTS" : ""
-        } ${this.tableName}`;
+          } ${this.tableName}`;
     }
   }
 
@@ -128,7 +113,7 @@ export class Table {
       default:
         return `CREATE TYPE ${enumCol.name} AS ENUM (${
           enumCol.columns.join(", ")
-        });`;
+          });`;
     }
   }
 
@@ -145,7 +130,7 @@ export class Table {
       case "sqlite3":
         return uniqueArray
           ? ` CREATE UNIQUE INDEX ${this.tableName}_${
-            uniqueArray.join("_")
+          uniqueArray.join("_")
           } ON ${this.tableName} (${uniqueString});`
           : "";
       case "mysql":
@@ -523,15 +508,15 @@ export class Table {
       this.dialect === "pg"
         ? typeName
         : this.dialect === "mysql"
-        ? "ENUM"
-        : "TEXT",
+          ? "ENUM"
+          : "TEXT",
       undefined,
       undefined,
       (col) =>
         this.dialect === "mysql" ? col.custom(`(${array.join(", ")})`)
-        : this.dialect === "sqlite3"
-        ? col.custom(`CHECK(${name} IN (${array.join(", ")}) )`)
-        : col,
+          : this.dialect === "sqlite3"
+            ? col.custom(`CHECK(${name} IN (${array.join(", ")}) )`)
+            : col,
     );
   }
